+++ ow-server-child-humidity.groovy
@@ -1,4 +1,4 @@
-def version() {'v0.1.5'}
+def version() {'v0.2.0'}
 
 metadata {
     definition (name: 'OW-Server 1-Wire - Child - Humidity',
@@ -7,12 +7,13 @@ metadata {
                 importUrl: 'https://raw.githubusercontent.com/ckamps/hubitat-drivers-ow-server/master/ow-server-child-humidity.groovy') {
         
         capability 'RelativeHumidityMeasurement'
-
+        capability 'TemperatureMeasurement'
         capability 'Refresh'
     }
 
     preferences {
-        input name: 'offset',    type: 'decimal', title: 'Humidity Offset',   description: '-n, +n or n to adjust sensor reading', range:'*..*'
+        input name: 'humidityOffset',    type: 'decimal', title: 'Humidity Offset',   description: '-n, +n or n to adjust sensor reading', range:'*..*'
+        input name: 'tempOffset',    type: 'decimal', title: 'Temperature Offset',   description: '-n, +n or n to adjust sensor reading', range:'*..*'
         input name: 'logEnable', type: 'bool',    title: 'Enable debug logging', defaultValue: false
     }
 }
@@ -35,11 +36,8 @@ def poll() {
 
 def refresh() {
     sensorId = device.deviceNetworkId
-    // Since AAG TAI-8540 sensors can have the same 1-Wire ID for both humidity and temp, by convention, we appended
-    // a trailing ".1" to the 1-Wire ID when we registered the humidity device.
-    sensorId = device.deviceNetworkId.substring(0, device.deviceNetworkId.length() - 2)
 
-    if (logEnable) log.debug("Getting humidity for sensor: ${sensorId}")
+    if (logEnable) log.debug("Getting humidity and temperature for sensor: ${sensorId}")
 
     try {
         humidity = getHumidity(sensorId)
@@ -50,7 +48,7 @@ def refresh() {
     }
 
     if (logEnable) log.debug("Humidity: ${humidity}")
-    humidity = offset ? (humidity + offset) : humidity
+    humidity = humidityOffset ? (humidity + humidityOffset) : humidity
     humidity = humidity.round(1)
 
     sendEvent(
@@ -58,6 +56,27 @@ def refresh() {
         value: humidity,
         unit: "%RH",
         descriptionText: "Humidity is ${humidity}%",
+    )
+
+    try {
+        temp = getTemperature(sensorId)
+    }
+    catch (Exception e) {
+        log.warn("Can't obtain temperature for sensor ${sensorId}: ${e}")
+        return
+    }
+    
+    if (logEnable) log.debug("Temperature - C: ${temp}")
+    temp = (location.temperatureScale == "F") ? ((temp * 1.8) + 32) : temp
+    temp = tempOffset ? (temp + tempOffset) : temp
+    temp = temp.round(2)
+
+    sendEvent(
+        name: 'temperature',
+        value: temp,
+        unit: "°${location.temperatureScale}",
+        descriptionText: "Temperature is ${temp}°${location.temperatureScale}",
+        translatable: true
     )
   
     def nowDay = new Date().format('MMM dd', location.timeZone)
@@ -84,4 +103,20 @@ private float getHumidity(sensorId) {
     if (!element.Humidity) throw new Exception("Humidity element does not exist in response from OW-Server for sensor ${sensorId}")
 
     return(element.Humidity.toFloat())
+}
+
+private float getTemperature(sensorId) {
+    def uri = "http://${parent.getOwServerAddress()}/details.xml"
+
+    response = parent.doHttpGet(uri)
+
+    if (!response) throw new Exception("doHttpGet to get temperature returned empty response ${sensorId}")
+
+    element = response.'**'.find{ it.ROMId == sensorId }
+    
+    if (!element) throw new Exception("Can't find matching ROMId element in response from OW-Server for sensor ${sensorId}")
+
+    if (!element.Temperature) throw new Exception("Temperature element does not exist in response from OW-Server for sensor ${sensorId}")
+
+    return(element.Temperature.toFloat())
 }