+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changes for Hubitat OW-Server Drivers
 
-## `ow-server-parent.groovy`
+### 0.2.0
+
+* Breaking change: No longer create separate devices for combined humidity and temperature sensors. Now create one device and set temperature and humidity settings for the one device.
+* Added support for DS2438-based temperature + humidity sensors.
+* Added support for EDS0065 OW-ENV-TH temperature + humidity sensor.
 
 ### 0.1